backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Felipe Sebben <felipesebben@yahoo.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.2"
uvicorn = {extras = ["standard"], version = "^0.36.0"}
sqlalchemy = "^2.0.43"
psycopg2-binary = "^2.9.10"
pydantic = "^2.11.9"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
